requirements.txt
anthropic
python-dotenv
mcp
mcp-cli
fastapi